Refactoring showModuleHierarchy.
